alter processing scores

+++ leaderboardstatapp/repository/migrations/1757697980_create_daily_contributor_scores_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS daily_contributor_scores (
     id SERIAL PRIMARY KEY,
     contributor_id BIGINT NOT NULL,
     user_id VARCHAR(100) NOT NULL,
-    daily_score BIGINT NOT NULL,
+    daily_score DOUBLE PRECISION NOT NULL
     -- TODO- add project_id
     rank BIGINT NOT NULL,
     timeframe VARCHAR(50) NOT NULL,
